components home,login,register moved in pages
