+++ src/App.tsx
@@ -2,15 +2,14 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Navbar from './components/Navbar/Navbar';
 import './App.scss';
-import Home from './components/Home/Home';
-import Register from './components/Register/Register';
-import Login from './components/Login/Login';
+import Home from './pages/Home/Home';
+import Register from './pages/Register/Register';
+import Login from './pages/Login/Login';
+import Footer from './components/Footer/Footer';
 
 const App = () => (
   <Router>
-    <div>
       <Navbar />
-      <main>
         <Switch>
           <Route path="/" exact>
             <Home />
@@ -22,8 +21,7 @@ const App = () => (
             Login
           </Route>
         </Switch>
-      </main>
-    </div>
+      <Footer />
   </Router>
     )
