+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Footer = () => (
+  <footer>
+    <p>
+      Made with
+      <span className="heart">&hearts;</span>
+      by Les Perruches
+    </p>
+</footer>
+  )
+
+export default Footer
