+++ src/pages/Home/Home.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Calendar from '../Calendar/Calendar'
+import Calendar from '../../components/Calendar/Calendar';
 
 const Home = () => (
   <>
