+++ src/sass/footer.scss
@@ -1,0 +1,15 @@
+$footer-height:4em;
+
+footer {
+  position: absolute; 
+  bottom: 0; 
+  left: 0; 
+  right: 0;
+  text-align: center;
+  background-color:white;
+  height:$footer-height;
+  width: 100%;
+  p {
+    color:$lightgray;
+  }
+}
